Refactor to abstract class
